Round 2 completed

+++ camper/views.py
@@ -5,7 +5,7 @@ from django.contrib.auth.decorators import login_required
 from .models import Camper
 
 # Camp registration fee
-registration_fee = 0
+registration_fee = 80
 
 
 
@@ -15,7 +15,7 @@ def home(request):
 
 
 # Camper list
-# @login_required(login_url='accounts:login')
+@login_required(login_url='accounts:login')
 def camper_list(request):
     camper_list = request.user.campers.all()
 
@@ -27,22 +27,15 @@ def camper_list(request):
 
 
 # Add new camper
-# @login_required(login_url='accounts:login')
+@login_required(login_url='accounts:login')
 def new_camper(request):
-    # Method 1
-    # camper = Camper(created_by=request.user)
     if request.method == 'POST':
-        # form = CamperForm(data=request.POST, instance=camper) # To support Method 1
         form = CamperForm(data=request.POST)
         if form.is_valid():
-            # Method 2
             camper = form.save(commit=False)
             camper.created_by = request.user
             camper.save()
-            # form.save()
-            # first_name = camper.cleaned_data.get('first_name') # Not working because now camper is an object of Camper
-            first_name = camper.first_name
-            messages.success(request, f"{first_name} has been successfully registered as a camper.")
+            messages.success(request, f"{camper.first_name} {camper.last_name} has been successfully registered as a camper.")
             return redirect('camper:list')
     else:
         form = CamperForm()
@@ -54,7 +47,7 @@ def new_camper(request):
 
 
 # Update an existing camper details
-# @login_required(login_url='accounts:login')
+@login_required(login_url='accounts:login')
 def update_camper(request, id):
     camper = get_object_or_404(Camper, pk=id)
     if camper.created_by != request.user:
@@ -65,7 +58,8 @@ def update_camper(request, id):
         if form.is_valid():
             form.save()
             first_name = form.cleaned_data.get('first_name')
-            messages.success(request, f"{first_name} has been successfully updated.")
+            last_name = form.cleaned_data.get('last_name')
+            messages.success(request, f"{first_name} {last_name} has been successfully updated.")
             return redirect('camper:list')
     else:
         form = CamperForm(instance=camper)
@@ -77,13 +71,17 @@ def update_camper(request, id):
 
 
 # Delete an existing camper
-# @login_required(login_url='accounts:login')
+@login_required(login_url='accounts:login')
 def delete_camper(request, id):
     camper = get_object_or_404(Camper, pk=id)
     if camper.created_by != request.user:
         messages.info(request, "You are not authorized to do this. This account has being logged for investigations", extra_tags='alert-danger')
         return redirect('camper:list')
     first_name = camper.first_name
-    camper.delete()
-    messages.info(request, f"{first_name} has been deleted as a camper.", extra_tags='alert-danger')
+    last_name = camper.last_name
+    if request.method == 'POST':
+        camper.delete()
+        messages.info(request, f"{first_name} has been deleted as a camper.", extra_tags='alert-danger')
+        return redirect('camper:list')
+    messages.info(request, f"You are not doing it rightly", extra_tags='alert-danger')
     return redirect('camper:list')
